add login/register

+++ server/models/todo.js
@@ -1,5 +1,6 @@
-const   mongoose = require('mongoose');
+const   mongoose = require('mongoose'),
         autoIncrement = require('mongoose-auto-increment');
+require('mongoose-double')(mongoose);
 autoIncrement.initialize(mongoose.connection);
 
 const TodoSchema = mongoose.Schema({
@@ -11,7 +12,10 @@ const TodoSchema = mongoose.Schema({
         default: 'Open'
     },
     order: {
-        type: Number,
+        type: Number
+    },
+    owner: {
+        type: String,
         index: true
     }
 });
@@ -20,29 +24,31 @@ TodoSchema.plugin(autoIncrement.plugin, { model: 'Todo', field: 'order' });
 
 const Todo = module.exports = mongoose.model('Todo', TodoSchema);
 
-module.exports.getTodos = (callback) => {
-    Todo.find({}, 'status text', {sort: {order: 1}}, callback);
+
+module.exports.getTodos = (owner, callback) => {
+    Todo.find({owner}, 'status text', {sort: {order: 1}}, callback);
 };
 
-module.exports.removeTodoById = (id, callback) => {
-    Todo.findOneAndRemove({_id: id}, callback);
+module.exports.removeTodoById = (owner, id, callback) => {
+    Todo.findOneAndRemove({owner, _id: id}, callback);
 };
 
-module.exports.addTodo = (text, callback) => {
+module.exports.addTodo = (owner, text, callback) => {
     const newTodo = Object.assign(new Todo(), {
-        text
+        text,
+        owner
     });
 
     newTodo.save(callback);
 };
 
-module.exports.changeTodoStatus = (id, status, callback) => {
-    Todo.findByIdAndUpdate(id, { status }, callback);
+module.exports.changeTodoStatus = (owner, id, status, callback) => {
+    Todo.findOneAndUpdate({owner, _id: id}, { status }, callback);
 };
 
 //not efficient
-module.exports.reorder = (newTodos, callback) => {
-    Todo.find({}, 'order', {sort: {order: 1}}, (err, todos) => {
+module.exports.reorder = (owner, newTodos, callback) => {
+    Todo.find({owner}, 'order', {sort: {order: 1}}, (err, todos) => {
         if(err) throw err;
 
         const bulkOps = [ ];
@@ -61,3 +67,29 @@ module.exports.reorder = (newTodos, callback) => {
     });
 };
 
+// module.exports.reorder = (todos, callback) => {
+//     let newOrder;
+//
+//     if(todos[1] && todos[2]){
+//         Todo.findOne({_id: todos[1]._id}, (err, firstTodo) => {
+//             if(err) throw err;
+//             Todo.findOne({_id: todos[2]._id}, (err, secondTodo) => {
+//                if(err) throw err;
+//                newOrder = firstTodo.order + (parseFloat((secondTodo.order - firstTodo.order).toFixed(4)) / 2);
+//                Todo.findByIdAndUpdate(todos[0]._id, { order: newOrder }, callback);
+//             });
+//         });
+//     } else if(todos[1]){       //new order is last
+//         Todo.findOne({_id: todos[1]._id}, (err, last) => {
+//             if(err) throw err;
+//             newOrder = last.order + 0.0005;
+//             Todo.findByIdAndUpdate(todos[0]._id, { order: newOrder }, callback);
+//         });
+//     } else {                    //new order is first
+//         Todo.findOne({_id: todos[2]._id}, (err, first) => {
+//             if(err) throw err;
+//             newOrder = parseFloat((first.order - 0.0005).toFixed(4));
+//             Todo.findByIdAndUpdate(todos[0]._id, { order: newOrder }, callback);
+//         });
+//     }
+// };
